Move clear_logs method to tests
