+++ tests/database/test_security.py
@@ -13,7 +13,7 @@ from server.database.security import (
     get_loaded_admins,
     number_of_admin_keys,
 )
-from server.logs import clear_logs
+from tests._utils.logs import clear_logs
 
 
 class Test_Getting_Admin(unittest.TestCase):
