+++ tests_integration/messages/test_timestamp.py
@@ -8,7 +8,7 @@ import server.app as _app
 from server.fleetv2_http_api.models.message import Message, Payload, DeviceId
 from server.enums import MessageType
 from server.database.time import timestamp
-from server.logs import clear_logs
+from tests._utils.logs import clear_logs
 
 
 class Test_Sending_And_Viewing_Command_For_Available_Car(unittest.TestCase):
